Add PhotoSizeSource FileType checks.
GitOrigin-RevId: 92685f8d05de9543e381ba753e8fe66dbeee286e

+++ td/telegram/files/FileType.h
@@ -14,7 +14,7 @@
 
 namespace td {
 
-enum class FileType : int8 {
+enum class FileType : int32 {
   Thumbnail,
   ProfilePhoto,
   Photo,
